Split out When_retry_has_been_sent into sub when's
