+++ src/ServiceControl.AcceptanceTests/ServiceControl.AcceptanceTests.csproj
@@ -234,12 +234,17 @@
     <Compile Include="MessageFailures\Is_System_Message_Tests.cs" />
     <Compile Include="MessageFailures\When_all_messages_are_retried.cs" />
     <Compile Include="MessageFailures\When_an_event_with_multiple_subscribers_fails.cs" />
+    <Compile Include="MessageFailures\When_a_failed_message_is_pending_retry.cs" />
+    <Compile Include="MessageFailures\When_a_pending_retry_is_resolved_by_queue_and_timeframe.cs" />
+    <Compile Include="MessageFailures\When_a_pending_retry_is_resolved_by_timeframe.cs" />
+    <Compile Include="MessageFailures\When_a_pending_retry_is_retried_again.cs" />
     <Compile Include="MessageFailures\When_a_multiple_error_message_with_same_uniqueid_are_imported_concurrently.cs" />
     <Compile Include="MessageFailures\When_an_error_message_has_ttbr_set.cs" />
     <Compile Include="MessageFailures\When_a_message_has_failed.cs" />
     <Compile Include="MessageFailures\When_a_message_has_failed_from_send_only_endpoint.cs" />
     <Compile Include="MessageFailures\When_a_invalid_id_is_sent_to_retry.cs" />
-    <Compile Include="MessageFailures\When_a_retry_has_been_sent.cs" />
+    <Compile Include="MessageFailures\When_a_pending_retry_is_retried_by_queue_and_timeframe.cs" />
+    <Compile Include="MessageFailures\When_a_pending_retry_is_resolved_by_selection.cs" />
     <Compile Include="MessageFailures\When_a_SagaComplete_message_fails.cs" />
     <Compile Include="MessageFailures\When_a_message_sent_from_third_party_endpoint_with_missing_metadata_failed.cs" />
     <Compile Include="MessageFailures\When_a_retry_fails_to_be_sent.cs" />
